(bugfix) Fix off-by-6 bug in assemble.rs
This fixes an off-by-six bug in assemble.rs that was causing
string-cache miss backup parsing to fail. When the string cache is
populated in `string_cache.rs`, from the given offset in the chunk
header, a `BinXmlNameLink` (which is 6 total bytes in size) is read from
the cursor, followed by the `BinXmlName`. On cache misses, `assemble.rs`
was not reading the `BinXmlNameLink` which was in turn failing to
advance the cursor by the number of bytes required.

This commit adds 6 to the offset in `assemble.rs`, fixing this problem.